jl040

+++ src/main/java/ch/ethz/idsc/owly/glc/core/OptimalAnyTrajectoryPlanner.java
@@ -134,6 +134,7 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
       System.out.println("node is already root");
       return 0;
     }
+    long tic = System.nanoTime();
     // needed for change of depthlimit
     int increasedDepthBy = newRoot.reCalculateDepth();
     // -- DEBUGING Values
@@ -191,42 +192,40 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
           final GlcNode nextParent = nextCandidatePair.getOrigin();
           // CandidateOrigin needs to be part of tree
           // if (!nextParent.isRoot()) { //TODO BUG? shoudl maybe be below?
-          if (!next.isRoot()) {
-            // check if Delete was properly conducted
-            if (deleteTreeCollection.contains(nextParent)) {
-              System.err.println("parent of node will be deleted --> not in QUEUE ");
-              throw new RuntimeException();
-            }
-            final List<StateTime> connector = //
-                getStateIntegrator().trajectory(nextParent.stateTime(), next.flow());
-            if (getObstacleQuery().isDisjoint(connector)) { // no collision
-              if (formerLabel.parent() != null)
-                formerLabel.parent().removeEdgeTo(formerLabel);
-              insertNodeInTree(nextParent, next);
-              candidateMap.get(domainKey).remove(nextCandidatePair);
-              addedNodesToQueue++;
-              if (!getGoalInterface().isDisjoint(connector))
-                offerDestination(next, connector);
-              break; // leaves the while loop, but not the for loop
-            }
+          // check if Delete was properly conducted
+          if (deleteTreeCollection.contains(nextParent)) {
+            System.err.println("parent of node will be deleted --> not in QUEUE ");
+            throw new RuntimeException();
+          }
+          final List<StateTime> connector = //
+              getStateIntegrator().trajectory(nextParent.stateTime(), next.flow());
+          if (getObstacleQuery().isDisjoint(connector)) { // no collision
+            if (formerLabel.parent() != null)
+              formerLabel.parent().removeEdgeTo(formerLabel);
+            insertNodeInTree(nextParent, next);
+            candidateMap.get(domainKey).remove(nextCandidatePair);
+            addedNodesToQueue++;
+            if (!getGoalInterface().isDisjoint(connector))
+              offerDestination(next, connector);
+            break; // leaves the while loop, but not the for loop
           }
         }
       }
     }
     System.out.println(addedNodesToQueue + " Nodes added to Domain = " + domainMap().size());
-    System.out.println("**Rootswitch finished**");
+    System.out.println("** Rootswitch finished in " + (System.nanoTime() - tic) * 1e-9 + "s **");
     return increasedDepthBy;
   }
 
-  // TODO JONAS Consistency checks
   @Override
   public void ObstacleUpdate(TrajectoryRegionQuery newObstacle) {
+    long tictotal = System.nanoTime();
     System.out.println("*** OBSTACLE UPDATE ***");
     long tic = System.nanoTime();
     setObstacleQuery(newObstacle);
     setBestNull();
     GlcNode root = getRoot();
-    // TODO JONAS: if root in collision
+    // TODO JONAS: What to do if root in collision
     List<GlcNode> domainMapList = new ArrayList<>(domainMap().values());
     Collections.sort(domainMapList, NodeDepthComparator.INSTANCE);
     int deletedNodes = 0;
@@ -244,7 +243,7 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
         }
         if (!label.isRoot()) {
           connector = getStateIntegrator().trajectory(label.parent().stateTime(), label.flow());
-          if (getObstacleQuery().isDisjoint(connector)) { // label not in Collision
+          if (getObstacleQuery().isDisjoint(connector)) { // label NOT in Collision
             while (!candidateQueue.isEmpty()) {
               // ******COPY PASTED //TODO JONAS in function
               CandidatePair nextBest = candidateQueue.element();
@@ -259,14 +258,12 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
                     if (subDeleteTree.size() > 1) // DEBUG INFO
                       System.err.println("Pruned Tree of Size: " + subDeleteTree.size());
                     CandidatePair formerLabelCandidate = new CandidatePair(label.parent(), label);
-                    label.parent().removeEdgeTo(label); // TODO confirm position or 5 lines below
+                    label.parent().removeEdgeTo(label);
                     if (!formerLabelCandidate.getCandidate().isLeaf()) {
                       System.err.println("The Candidate in the bucket has children");
                       throw new RuntimeException();
                     }
                     getCandidateMap().get(domainKey).add(formerLabelCandidate);
-                    // formerLabel disconnecting
-                    // if (label.parent() != null) //TODO confirm if not needed
                     // adding next to tree and DomainMap
                     insertNodeInTree(nextBest.getOrigin(), nextBestNode);
                     addedNodes++;
@@ -286,7 +283,7 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
               if (removedCP != nextBest)
                 throw new RuntimeException();
             } // else do nothing
-          } else { // DELETE label, as in collision
+          } else { // DELETE label, as IN collision
             deletedNodes++;
             final Collection<GlcNode> subDeleteTree = deleteSubtreeOf(label);
             if (subDeleteTree.size() > 1) // DEBUG INFO
@@ -386,7 +383,10 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
         + " domains took :" + (toc - tic) * 1e-9 + "s");
     System.out.println("Nodes added:  " + addedNodes);
     System.out.println("Nodes deleted:" + deletedNodes);
-    System.out.println("**ObstacleUpdate finished**");
+    System.out.println("** ObstacleUpdate finished in " + (System.nanoTime() - tictotal) * 1e-9 + "s **");// Consistency checks
+    DebugUtils.nodeAmountCompare(this);
+    DebugUtils.connectivityCheck(domainMap().values());
+    DebugUtils.heuristicConsistencyCheck(this);
   }
 
   public Map<Tensor, Set<CandidatePair>> getCandidateMap() {
@@ -406,6 +406,7 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
   /* package */ final void RelabelingDomains() {
     GlcNode root = getRoot();
     List<GlcNode> treeList = new ArrayList<GlcNode>(Nodes.ofSubtree(root));
+    System.err.println("checking for domainlabel changes due to heuristic change,  Treesize: " + treeList.size());
     treeList.stream().parallel() //
         .forEach(glcNode -> glcNode.setMinCostToGoal(getGoalInterface().minCostToGoal(glcNode.state())));
     Collections.sort(treeList, NodeDepthComparator.INSTANCE);
@@ -414,31 +415,35 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
     for (GlcNode label : treeList) {
       // iterating through all Nodes in Tree starting at lowest depth
       Tensor domainKey = convertToKey(label.state());
-      if (getNode(domainKey) == label) { // this node could have been deleted from tree in prev iteration
-        // if (!getNode(domainKey).equals(current)) {
-        // throw new RuntimeException(); // current should be labeling its own domain
-        // TODO JONAS: skip loop instead of runtime exception,
-        // could be relabeled in loopiteration before, therefore no check is needed
+      if (getNode(domainKey) == label && !label.isRoot()) { // this node could have been deleted from tree in prev iteration
+        // could be relabeled in loopiteration before, therefore no check is needed if changed
         if (getCandidateMap().containsKey(domainKey)) {
           Set<CandidatePair> tempCandidateSet = candidateMap.get(domainKey);
           tempCandidateSet.parallelStream().forEach(cp -> //
           cp.getCandidate().setMinCostToGoal(getGoalInterface().minCostToGoal(cp.getCandidate().state())));
           PriorityQueue<CandidatePair> candidateQueue = new PriorityQueue<>(tempCandidateSet);
           while (!candidateQueue.isEmpty()) {
-            final CandidatePair nextCandidatePair = candidateQueue.poll();
+            final CandidatePair nextCandidatePair = candidateQueue.element();
             final GlcNode possibleCandidateNode = nextCandidatePair.getCandidate();
             if (Scalars.lessThan(possibleCandidateNode.merit(), label.merit())) {
               // collision check only if new node is better
               final GlcNode possibleCandidateOrigin = nextCandidatePair.getOrigin();
-              if (Nodes.listFromRoot(possibleCandidateOrigin).get(0) == root) {
+              if (Nodes.areConnected(possibleCandidateOrigin, root)) {
                 final List<StateTime> connector = //
                     getStateIntegrator().trajectory(possibleCandidateOrigin.stateTime(), possibleCandidateNode.flow());
                 if (getObstacleQuery().isDisjoint(connector)) {
                   Collection<GlcNode> deleteTree = deleteSubtreeOf(label);
+                  if (deleteTree.size() > 1) // DEBUG INFO
+                    System.err.println("Pruned Tree of Size: " + deleteTree.size());
+                  CandidatePair formerLabelCandidate = new CandidatePair(label.parent(), label);
+                  label.parent().removeEdgeTo(label);
+                  if (!formerLabelCandidate.getCandidate().isLeaf()) {
+                    System.err.println("The Candidate in the bucket has children");
+                    throw new RuntimeException();
+                  }
+                  getCandidateMap().get(domainKey).add(formerLabelCandidate);
                   deletedNodes = deletedNodes + deleteTree.size() - 1; // -1 as this node was replaced
                   replacedNodes++;
-                  if (label.parent() != null)
-                    label.parent().removeEdgeTo(label);
                   insertNodeInTree(possibleCandidateOrigin, possibleCandidateNode);
                   candidateMap.get(domainKey).remove(nextCandidatePair);
                   if (!getGoalInterface().isDisjoint(connector))
@@ -447,50 +452,54 @@ public class OptimalAnyTrajectoryPlanner extends AbstractAnyTrajectoryPlanner {
                 }
               } else { // CandidateOrigin is not connected to tree anymore --> Remove this Candidate from Map
                 candidateMap.get(convertToKey(possibleCandidateNode.state())).remove(nextCandidatePair);
-                final CandidatePair removedCP = candidateQueue.remove();
-                if (removedCP != possibleCandidateNode)
-                  throw new RuntimeException(); // removed candidate should be nextBest from before
               }
             } else {
               break; // if no better Candidates are found leave while of Candidates loop -> Speedgain
             }
+            final CandidatePair removedCP = candidateQueue.remove();
+            if (removedCP != nextCandidatePair)
+              throw new RuntimeException(); // removed candidate should be nextBest from before
           }
         }
       }
     }
     long candidateMapBeforeSize = candidateMap.size();
     long oldtotalCandidates = candidateMap.values().parallelStream().flatMap(Collection::stream).count();
-    System.out.println("replaced Nodes: " + replacedNodes + " deleted Nodes: " + deletedNodes + ", during relabel");// stopped iterating over tree
     candidateMap.entrySet().parallelStream().forEach(candidateSet -> candidateSet.getValue()//
         .removeIf(cp -> !Nodes.rootFrom(cp.getOrigin()).equals(root)));
     // Deleting CandidateBuckets, if they are empty
     candidateMap.values().removeIf(bucket -> bucket.isEmpty());
     long newtotalCandidates = candidateMap.values().parallelStream().flatMap(Collection::stream).count();
+    // INFO
+    System.out.println("replaced Nodes: " + replacedNodes + " deleted Nodes: " + deletedNodes + ", during relabel");// stopped iterating over tree
     System.out.println(oldtotalCandidates - newtotalCandidates + " of " + oldtotalCandidates + //
         " C. removed from CL. with Origin in deleteTree: " + newtotalCandidates);
     System.out.println("CandidateMap before " + candidateMapBeforeSize + //
         " and after: " + candidateMap.size());
+    DebugUtils.nodeAmountCompare(this);
   }
 
   @Override
   protected final boolean GoalCheckTree(final Region possibleGoalNodesRegion) {
     final Collection<GlcNode> treeCollection = Nodes.ofSubtree(getRoot());
-    // Smart way: uses 15% -> 30% of the time of normal implemenation
-    // Smart way: uses 30% -> 40% of the time of paralel implementation
+    // Smart way: uses 15% -> 30% of the time of normal implementation
+    // Smart way: uses 30% -> 40% of the time of parallel implementation
     // , tested with R2GlcAnyCircleDemo
-    List<GlcNode> possibleGoalNodes = new ArrayList<>();
+    Collection<GlcNode> possibleGoalNodes = new ArrayList<>();
     possibleGoalNodes = treeCollection.stream().parallel()//
         .filter(node -> possibleGoalNodesRegion.isMember(node.state()))//
         .collect(Collectors.toList());
     // checking only Nodes, which could reach the Goal
     System.out.println("Total Nodes in Tree: " + treeCollection.size() + " possibleGoalNodes: " + possibleGoalNodes.size());
+    // --
     return GoalCheckTree(possibleGoalNodes);
   }
 
   private final boolean GoalCheckTree(Collection<GlcNode> treeCollection) {
     // Parallel: 15%-50% Speedgain, tested with R2GlcConstTimeHeuristicAnyDemo,
     // TODO why does parallel give different result? then non parallel? e.g. R2GlcAnyCircleDemo
-    // TODO JONAS try again, i added 'synchronized' to offerDestination
+    //
+    // TODO JAN: No difference with synchronized, same bug, even with exactly the same List
     treeCollection.stream().forEach(node -> {
       if (!node.isRoot()) {
         final List<StateTime> trajectory = getStateIntegrator().trajectory(node.parent().stateTime(), node.flow());
